requirements.txt
sanic==22.6.2
sanic-validation==0.5.1
sanic-session==0.8.0
SQLAlchemy==1.4.40
databases==0.6.1
alembic==1.8.1
psycopg2-binary==2.9.3
asyncpg==0.26.0
redis==4.3.4
PyJWT==1.7.1
pycryptodome==3.15.0
python-dotenv==0.20.0
manage.py==0.2.10
flake8==5.0.4
flake8-isort==4.2.0
black==22.6.0

git+https://github.com/AkbarYusupov2003/sanic_openapi_distutils_deprication_fix.git@master
